requirements.txt
requests
django-cors-headers
djangorestframework
Django
redis